organize non-context-dependent validation methods in dedicated object
